PDU unit/slot

+++ site/ajax_pdu.py
@@ -27,19 +27,19 @@ def pdu_list_units(aWeb):
 
  db = DB()
 
- optemplate = "<A CLASS='z-btn z-small-btn z-btnop' OP=load SPIN=true DIV=div_navleft LNK='ajax.cgi?call=pdu_list_units&pdu={0}&nstate={1}&slot={2}'><IMG SRC='images/btn-{3}'></A>"
+ optemplate = "<A CLASS='z-btn z-small-btn z-btnop' OP=load SPIN=true DIV=div_navleft LNK='ajax.cgi?call=pdu_list_units&pdu={0}&nstate={1}&slot={2}&unit={3}'><IMG SRC='images/btn-{4}'></A>"
 
  if len(pdulist) == 0:
   pdulist.append(pduop)
  print "<DIV CLASS='z-framed'><DIV CLASS='z-table'><TABLE WIDTH=330>"
  print "<TR><TH>PDU</TH><TH>Entry</TH><TH>Device</TH><TH style='width:63px;'>State</TH></TR>"
  for pdu in pdulist:
-  avocent = Avocent(pdu,domain)
+  avocent = Avocent(pdu)
   avocent.load_snmp()
 
   # Ops
-  if pdu == pduop and slot and nstate:
-   avocent.set_state(slot,nstate)
+  if pdu == pduop and slot and unit and nstate:
+   avocent.set_state(slot,unit,nstate)
    # Avocent is not fast enough to execute something immediately after reboot op, halt output then :-)
    if nstate == 'reboot':
     from time import sleep
@@ -47,30 +47,31 @@ def pdu_list_units(aWeb):
 
   for key in avocent.get_keys(aSortKey = lambda x: int(x.split('.')[0])*100+int(x.split('.')[1])):
    value = avocent.get_entry(key)
-   print "<TR><TD TITLE='Open up a browser tab for {1}'><A TARGET='_blank' HREF='https://{0}:3502'>{1}</A></TD><TD>{2}</TD>".format(avocent._ip,pdu,value['pduslot'])
-   print "<TD><A CLASS='z-btnop' OP=load DIV=div_navcont LNK='ajax.cgi?call=pdu_unit_info&domain={0}&pdu={1}&slot={2}&name={3}&slotname={4}' TITLE='Edit port info' >{3}</A></TD><TD>".format(domain,pdu,key,value['name'], value['pduslot'])
+   print "<TR><TD TITLE='Open up a browser tab for {1}'><A TARGET='_blank' HREF='https://{0}:3502'>{1}</A></TD><TD>{2}</TD>".format(avocent._ip,pdu,value['slotname']+'.'+value['unit'])
+   print "<TD><A CLASS='z-btnop' OP=load DIV=div_navcont LNK='ajax.cgi?call=pdu_unit_info&pdu={0}&slot={1}&unit={2}&name={3}&slotname={4}' TITLE='Edit port info' >{3}</A></TD><TD>".format(pdu,value['slot'],value['unit'],value['name'], value['slotname'])
    if value['state'] == "off":
-    print optemplate.format(pdu, "on", key, "start")
+    print optemplate.format(pdu, "on", value['slot'],value['unit'], "start")
    else:
-    print optemplate.format(pdu, "off", key, "shutdown")
-    print optemplate.format(pdu, "reboot", key, "reboot")
+    print optemplate.format(pdu, "off", value['slot'],value['unit'], "shutdown")
+    print optemplate.format(pdu, "reboot", value['slot'],value['unit'], "reboot")
    print "</TD></TR>"
  print "</TABLE></DIV></DIV>"
 
 def pdu_unit_info(aWeb):
  pdu  = aWeb.get_value('pdu')
  slot = aWeb.get_value('slot')
+ unit = aWeb.get_value('unit')
  slotname = aWeb.get_value('slotname')
  name = aWeb.get_value('name')
- domain = aWeb.get_value('domain')
  print "<DIV CLASS='z-framed z-table' style='resize: horizontal; margin-left:0px; width:420px; z-index:101; height:150px;'>"
  print "<FORM ID=pdu_form>"
  print "<INPUT NAME=slot   VALUE={} TYPE=HIDDEN>".format(slot)
+ print "<INPUT NAME=unit   VALUE={} TYPE=HIDDEN>".format(unit)
  print "<INPUT NAME=pdu    VALUE={} TYPE=HIDDEN>".format(pdu)
  print "<TABLE style='width:100%'>"
  print "<TR><TH COLSPAN=2>PDU Slot Info</TH></TR>"
  print "<TR><TD>PDU:</TD><TD>{0}</TD></TR>".format(pdu)
- print "<TR><TD>Slot.Unit:</TD><TD>{0}</TD></TR>".format(slotname)
+ print "<TR><TD>Slot.Unit:</TD><TD>{0}.{1}</TD></TR>".format(slotname,unit)
  print "<TR><TD>Name:</TD><TD><INPUT NAME=name TYPE=TEXT CLASS='z-input' PLACEHOLDER='{0}'></TD></TR>".format(name)
  print "<TR><TD COLSPAN=2>&nbsp;</TD></TR>"
  print "</TABLE>"
@@ -191,8 +192,9 @@ def pdu_unit_update(aWeb):
   name = aWeb.get_value('name')
   pdu  = aWeb.get_value('pdu')
   slot = aWeb.get_value('slot')
+  unit = aWeb.get_value('unit')
   avocent = Avocent(pdu)
-  avocent.set_name(slot,name)
+  avocent.set_name(slot,unit,name)
   print "Updated name: {} for {} slot {}".format(name,pdu,slot)
  else:
   print "Name not updated"
